Modularize rules_sh
For compatibility with bzlmod and addition to the
https://github.com/bazelbuild/bazel-central-registry.

Adds a module rule `sh_configure` for the POSIX toolchain configured by
`sh_posix_configure`. Module rules may invoke repository rules, but may
not register toolchains. This commit adds a boolean parameter `register`
to `sh_posix_configure` for compatibility with both the old repository
rule workflow and the new module workflow.