Avoid runfiles dependency for true/false
sh_binary for a Shell script creates a wrapper binary that relies on the
runfiles mechanism to discover the corresponding shell script at
runtime. This introduces an implicit dependency on the runfiles
mechanism.

The purpose of the true/false toolchain is to test sh_binaries that
don't depend on runfiles, therefore the shell script approach is not
appropriate.

This uses C++ binaries instead.