Use rlocationpath instead of rootpath
Bazel 8 uses the plus sign instead of tilde in repo names now (the
`--incompatible_use_plus_in_repo_names` flag is a no-op).

That causes the `rlocation` lookup for the arguments of the
`@rules_sh_tests//import:create_shim_test` to fail and just return empty
results.

```
comm: '': No such file or directory
Expected files to be equal, but they are different: '' and ''.
```